Remove calling notifyChildInserted in executeParserNodeInsertionIntoIsolatedTreeWithoutNotifyingParent in ContainerNodeAlgorithms.cpp https://bugs.webkit.org/show_bug.cgi?id=288982 rdar://146035398
Reviewed by Ryosuke Niwa.

This PR is put up currently to test correctness on Layout Tests.

When parsing and constructing a temporary DocumentFragment using the HTMLFastPathParser,
we insert un-connected nodes and update the node flags of this node's subtree by calling
notifyChildInserted via executeParserNodeInsertionIntoIsolatedTreeWithoutNotifyingParent.
notifyChildInserted does a traversal of the inserted node's subtree to do this updating.
When appending the constructed DocumentFragment's children to be children of the ContainerNode
that the user has called .innerHTML on, we also call notifyChildInserted on the each of these
fragment nodes from appendChildWithoutPreInsertionValidityCheck via replaceChildrenWithFragment.
This visits and updates the node flags of every node in the DocumentFragment tree.
Skip the traversals to update the node flags when inserting nodes into the DocumentFragment,
since we always finally update the flags of the inserted nodes when they are inserted into
the permanent Document. This is a .64% progression on SP3.

This PR was influenced by this Chromium commit: https://chromium-review.googlesource.com/c/chromium/src/+/6174392

* Source/WebCore/dom/ContainerNode.cpp:
(WebCore::executeParserNodeInsertionIntoIsolatedTreeWithoutNotifyingParent):

Canonical link: https://commits.webkit.org/291710@main