[IndexedDB] IDBBackingStore::revertAddIndex should only be invoked after IDBBackingStore::addIndex https://bugs.webkit.org/show_bug.cgi?id=289288 rdar://146071686
Reviewed by Basuke Suzuki.

When IDBBackingStore::addIndex is not invoked or when it returns error, it means the index is not added successfully to
the backing store. In this case, we should not invoke IDBBackingStore::revertAddIndex, as it can lead to a previously
added index gets removed.

* Source/WebCore/Modules/indexeddb/server/UniqueIDBDatabase.cpp:
(WebCore::IDBServer::UniqueIDBDatabase::createIndexAsyncAfterQuotaCheck):
(WebCore::IDBServer::UniqueIDBDatabase::didCreateIndexAsyncForTransaction):
* Source/WebCore/Modules/indexeddb/server/UniqueIDBDatabase.h:
* Source/WebCore/Modules/indexeddb/server/UniqueIDBDatabaseTransaction.cpp:
(WebCore::IDBServer::UniqueIDBDatabaseTransaction::deleteIndex):
(WebCore::IDBServer::UniqueIDBDatabaseTransaction::createIndex):

Canonical link: https://commits.webkit.org/291790@main