[iOS] [Visual Bidi Selection] Selection flickers when selecting LTR text next to Indic numerals https://bugs.webkit.org/show_bug.cgi?id=289302 rdar://144607260
Reviewed by Aditya Keerthi.

Consider the following scenario, where two bidi LTR text runs are embedded in an RTL paragraph, with
only a space character in between; let's suppose the entire left word (second in logical order) is
selected below, with `S` and `E` denoting the selection start and end, respectively. The selection
highlight (as seen by the user) is also denoted below.

        "Hello"      "۷۴۸۵۵"
      S        E
    | ▦▦▦▦▦▦▦▦▦▦ | ▦▦▦▦▦▦▦▦▦▦ |
    | ---ltr---> | ---ltr---> |
    |<------------------------| rtl ¶

      ▦▦▦▦▦▦▦▦▦▦               selection highlight

Now, suppose the user takes the selection end handle (`E`) and extends it into the right word, such
that the selection handles flip (due to reversing the logical base/extent) — that is, the previous
selection end becomes the new start of the selection, and the previous start becomes the new end:

        "Hello"      "۷۴۸۵۵"
               E        S
    | ▦▦▦▦▦▦▦▦▦▦ | ▦▦▦▦▦▦▦▦▦▦ |
    | ---ltr---> | ---ltr---> |
    |<------------------------| rtl ¶

                ▦▦▦▦▦▦▦▦▦       selection highlight

Currently, our visual bidi selection heuristics naively join the two logically-contiguous, visually-
discontiguous selections together by combining them together to form a single, coalesced rect that
spans from the start position to the end position. In this particular scenario, this means that the
left word ("Hello" in this example) will appear visually unselected even though it's still selected
in the DOM, which in turn creates a confusing experience for the user (especially since it will be
selected both logically and visually once the user lets go anyways).

To fix this, we augment the heuristics in `makeBidiSelectionVisuallyContiguousIfNeeded`:

-   First, detect this above scenario by checking for a single-line selection, wherein all selection
    rects are of an opposite direction as the paragraph (or primary direction of the single-line
    selection range). In the above example, this would be two or more adjacent LTR selections, in an
    RTL paragraph.

-   Use the opposite endpoints of the selection rect when determining the final bounding rect of the
    coalesced, visually-contiguous selection. In the above example, instead of using the endpoints
    at the end of the selected text, we'd use the start endpoints instead. To determine whether the
    start or end of the selection needs this treatment, we use `RenderedPosition` to check whether
    the start or end position lies on a bidi visual boundary.

The end result is that the selection will visually span from the start to the left-most visual
extent of "Hello" while the user is extending the selection:

        "Hello"      "۷۴۸۵۵"
               E        S
    | ▦▦▦▦▦▦▦▦▦▦ | ▦▦▦▦▦▦▦▦▦▦ |
    | ---ltr---> | ---ltr---> |
    |<------------------------| rtl ¶

      ▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦       selection highlight

...and it will then snap to select the entire paragraph after the user lets go:

      "Hello"      "۷۴۸۵۵"
               E            S
    | ▦▦▦▦▦▦▦▦▦▦ | ▦▦▦▦▦▦▦▦▦▦ |
    | ---ltr---> | ---ltr---> |
    |<------------------------| rtl ¶

      ▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦▦   selection highlight

* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-7.html:
* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-8.html:
* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-9-expected.txt: Added.
* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-9.html: Added.

Add a new layout test to exercise the change, by verifying that selected text does not get visually
unselected, when long pressing on "Hello" and extending the selection end handle into Urdu (Indic)
numerals. The final logically and visually-contiguous selection should encompass the entire
paragraph.

* LayoutTests/resources/ui-helper.js:
(window.UIHelper.computeLineBounds):
* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::computeSelectionEndpointDirections):
(WebCore::makeBidiSelectionVisuallyContiguousIfNeeded):

See above for more details.

(WebCore::adjustTextDirectionForCoalescedGeometries):
(WebCore::RenderObject::collectSelectionGeometries):

Canonical link: https://commits.webkit.org/291799@main