[macOS] MiniBrowser occasionally launches with a giant top inset when Screen Time is enabled https://bugs.webkit.org/show_bug.cgi?id=289350 rdar://145900121
Reviewed by Wenson Hsieh.

As a preface, all use of the term "constraints" in this commit message refers
to "user" constraints, and not autoresizing mask constraints.

`WKWebView` automatically computes a top content inset using the content layout
rect of its window and converting to its own coordinate space. In the common
case, this inset ends up being the height of the window's titlebar.

The logic to automatically compute this inset has existed since 152497@main, and
is performed by: scheduling an update, waiting a runloop turn, and finally,
computing the inset.

Updates are scheduled in three cases:
- `WKWebView` initialization
- `-[WKWebView renewGState]`, which generally corresponds to frame changes
- Whenever the containing window's `contentLayoutRect` changes, using KVO

MiniBrowser's window style results in the creation of an `NSThemeFrame`, which
is a direct descendant of the `NSWindow`. The `NSThemeFrame`'s
`contentLayoutGuide`'s frame is the window's `contentLayoutRect`. This is what
is used to compute the top content inset. The `NSThemeFrame` is henceforth
referred to as the "border view".

With the introduction of Screen Time, an `STWebpageView` is added as a subview
of `WKWebView`, in the first layer tree commit after the first committed load.
`STWebpageView` has an `NSRemoteView` subview, that is added asynchronously
after the `STWebpageView` itself is parented. The remote view is only added
once a reply is received from Screen Time's view service.

In MiniBrowser, the added remote view is the first view to use Auto Layout. Once
the border view's view hierarchy contains a view using Auto Layout, AppKit
updates the border view's layout engine. This update triggers an update to
the border view's content layout guide's constraints, and AppKit uses KVO to
notify the observer (`WebViewImpl`) of a change to the window's
`contentLayoutRect`. In response to the content layout rect change, WebKit
schedules a top constent inset update, as explained above.

Now at this point, there only exists constraints on the border view's content
layout guide, and the remote view. Importantly, there are not yet any
constraints on the border view itself. The constraints on the border view are
added in the next display link flush. This means that both of the following
scenarios are possible:

Scenario A (good)
1. Top content inset update scheduled due to introduction of Auto Layout.
2. Border view constraints added on display link flush.
3. Top content inset update dispatched on runloop turn.

Scenario B (bad)
1. Top content inset update scheduled due to introduction of Auto Layout.
2. Top content inset update dispatched on runloop turn.
3. Border view constraints added on display link flush.

Scenario A is the "good case". All the relevant constraints (most importantly,
the border view's constraints) are in their "final" state, and the correct
content layout rect is observed at the time of top content inset update dispatch.
On the other hand, Scenario B represents the issue: the inset update occurs
with incomplete constraints, ultimately resulting in a top content inset equal
to the height of the window. Additionally, when the border view constraints are
finally added in Scenario B, the border's view frame and content layout guide
are updated, but AppKit does not report the change via KVO, leaving `WKWebView`
stuck in state. This lack of update is a clear AppKit bug, whereas the timing
issue is less obviously so.

Consequently, the issue can be fixed by ensuring that border view constraints
are present at the time of top content inset update dispatch. This is achieved
by simply calling `-[NSWindow updateConstraintsIfNeeded]`.

Thanks to Jessica Cheung for assisting with the investigation of this issue.

* Source/WebKit/UIProcess/mac/PageClientImplMac.mm:
(WebKit::PageClientImpl::computeAutomaticTopObscuredInset):

Ensure the window's and border view's constraints are up-to-date prior to
accessing the window's content layout rect. Without this change, there is a
chance that the content layout rect is computed with an incomplete set of
constraints, resulting in an incorrect value, and AppKit failing to notify
clients of an eventually correct value.

This call is a no-op is Auto Layout is not being used.

Canonical link: https://commits.webkit.org/291818@main