Add support for generating more pair consumers https://bugs.webkit.org/show_bug.cgi?id=289432
Reviewed by Tim Nguyen.

Switch more consumers that use the form `[...]{1,2}` in their
grammars to match the modeling used already where a coalescing
CSSValuePair is used. With the consumer generated, the style
builder and computed style extractors need to be updated to
match.

These allow generation of:
    -webkit-initial-letter
    scroll-snap-align

* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/ComputedStyleExtractor.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Inline.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Inline.h:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+ScrollSnap.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+ScrollSnap.h:
* Source/WebCore/style/StyleBuilderConverter.h:

Canonical link: https://commits.webkit.org/291873@main