[WTF] Separate SequesteredAllocator::Arena logic from parent https://bugs.webkit.org/show_bug.cgi?id=289491 rdar://146690839
Reviewed by Yusuke Suzuki.

Previously, the Arena had to hardcode layout information from
SequesteredArenaAllocator so that it could find its parent allocator.
This is safe because we know that each Arena only ever exists within
a SequesteredArenaAllocator, but the logic properly belongs inside the
SAA itself so this patch moves it there.

Symmetrically, the SequesteredArenaAllocator was handling logic that
rather belongs inside of the Arena itself. Putting this there will
enable the usage of multiple separate Arenas w/in a single allocator
in the future.

* Source/WTF/wtf/SequesteredAllocator.h:

Canonical link: https://commits.webkit.org/291969@main