REGRESSION (Safari 18.2): Allow < 12 byte IV for encrypt rdar://146395205 https://bugs.webkit.org/show_bug.cgi?id=289171
Reviewed by Pascoe.

As part of CryptoKit adoption we did add a limit to the minimum
IV size for AES-GCM encrypt. This change removes that and falls back to the CommonCrypto
Implementation if IV is < 12 bytes.

A Test has been added to verify the same.

* LayoutTests/crypto/subtle/aes-gcm-generate-key-encrypt-decrypt-small-iv-expected.txt: Added.
* LayoutTests/crypto/subtle/aes-gcm-generate-key-encrypt-decrypt-small-iv.html: Added.
* Source/WebCore/crypto/cocoa/CryptoAlgorithmAESGCMMac.cpp:
(WebCore::encryptCryptoKitAESGCM):
(WebCore::CryptoAlgorithmAESGCM::platformEncrypt):

Canonical link: https://commits.webkit.org/291981@main