[ResourceMonitor] Cannot go back to previous page by Back button after unloading of iframe. https://bugs.webkit.org/show_bug.cgi?id=289623 rdar://146671361
Reviewed by Brady Eidson.

When ResourceMonitor unloads iframe, a placeholder is displayed using `srcdoc`. Setting `srcdoc` is
a navigation so this action is pushed onto the history. This is not expected behavior.

Adding optional parameter to set the visibility of the navigation in the history when setting `srcdoc`.
The visibility is stored in iframe element and used when SubstituteData is created.

* Source/WebCore/html/HTMLIFrameElement.cpp:
(WebCore::HTMLIFrameElement::setSrcdoc):
* Source/WebCore/html/HTMLIFrameElement.h:
* Source/WebCore/loader/FrameLoader.cpp:
(WebCore::FrameLoader::defaultSubstituteDataForURL):
* Source/WebCore/page/LocalFrame.cpp:
(WebCore::LocalFrame::showResourceMonitoringError):

Canonical link: https://commits.webkit.org/292053@main